Extract ChatPanel browser test environment
